+++ src/Vehistra.Client/Views/DriverDetailView.xaml.cs
@@ -1,0 +1,12 @@
+using System.Windows.Controls;
+
+namespace Vehistra.Client.Views;
+
+/// <summary>Fahrerakte mit Stammdaten, Fahrzeugen, Schaeden, Unfaellen und Werkstatt.</summary>
+public partial class DriverDetailView : UserControl
+{
+    public DriverDetailView()
+    {
+        InitializeComponent();
+    }
+}